# Performance issue in check_tls_padding