@randombit 2017-09-29
Thanks for the profiling work. Certainly there is no reason to iterate over the entire record here. Your patch is I think not quite right, because what it ends up doing is checking the first 256 bytes of the record, but if the record is longer than that, it will accept any (even invalid) padding. There was no existing test that would catch this, however. I just pushed a new test that demonstrates the issue.