@randombit 2017-09-29
I just addressed this issue in 163c5b4a3, so closing this PR. If you can rerun your profiling to confirm that would be lovely.

Again thank you for profiling and raising these issues, it is very useful.