@ma30002000 2017-10-04
Thanks a lot for addressing the issue that quickly and adding the useful test case. I can confirm that on 2.3.0, the mentioned performance issue is resolved.